Automatically backed up by Learn

+++ db/migrate/003_create_haunted_houses.rb
@@ -4,7 +4,7 @@ class CreateHauntedHouses < ActiveRecord::Migration[5.1]
       t.string :name
       t.string :location
       t.string :theme
-      t.integer :num_of_employees
+      t.float :num_of_employees
       t.boolean :still_in_business
       t.timestamp :opening_time
       t.timestamp :closing_time
